+++ packages/rescript-kaplay/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+### Changed
+- `Context.loop` is now `Context.loopWithController`
+
+### Added
+- `Context.wait`
+
 ## [0.10.1] - 2025-11-23
 ### Changed
 - Update to ReScript 12 RC